backend/requirements.txt
alembic==1.17.2
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
